Various fixes
Change to .travis-ucl.sh is to build support for signatures into the test machine's copy of libucl.
Change to decoder_test.go is to show what is expected in random failing test (see issue #6).
Change to object.go is to remove memory leak, and make sure created C strings are freed.